Refactor: Update import statements to replace references from _etrace_operators to _etrace_op across multiple files
